+++ src/main/java/com/black_dog20/permissionlevels/PermissionLevels.java
@@ -1,5 +1,6 @@
 package com.black_dog20.permissionlevels;
 
+import com.black_dog20.permissionlevels.commands.CommandBypassPlayerLimit;
 import com.black_dog20.permissionlevels.commands.CommandOp;
 import com.google.common.collect.ImmutableSet;
 import net.minecraft.commands.Commands;
@@ -32,6 +33,7 @@ public class PermissionLevels
     public void onCommandRegister(RegisterCommandsEvent event) {
         if (environmentTypeSet.contains(event.getCommandSelection())) {
             CommandOp.register(event.getDispatcher());
+            CommandBypassPlayerLimit.register(event.getDispatcher());
         }
     }