+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Permission Levels Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [1.19.1, 1.0.4] - 2022-08-14
+- Added option to set `bypassPlayerLimit` when using the `op` or `xop` command.
+- Added another command `bypassPlayerLimit` that directly sets if an op is allowed to bypass the player limit.
+
 ## [1.19, 1.0.3] - 2022-07-31
 - Updated to Forge version 41.1.0 to work with the new client refactor and rename refactor